catalog: persist all ID allocations
Global IDs could previously get reused after a crash, because upon
reboot the ID allocator was reset to one greater than the maximum ID
present in the catalog. This was once safe, because any state that
referred to those reused IDs would have evaporated during the crash, but
causes trouble for the timestamp persistence code, which might persist
messages referring to those recycled IDs.

Instead, have every ID allocation durably record the new state of the ID
allocator, so that we can guarantee IDs are not reused even after a
reboot. This has potential performance implications, as there is now an
fsync on the path for slow-path peeks (not fast-path peeks, though!),
but there are bigger fish to fry with slow-path peeks anyway.

+++ src/expr/id.rs
@@ -38,12 +38,12 @@ pub trait IdHumanizer {
 
 /// The identifier for a local component of a dataflow.
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
-pub struct LocalId(usize);
+pub struct LocalId(u64);
 
 impl LocalId {
     /// Constructs a new local identifier. It is the caller's responsibility
     /// to provide a unique `v`.
-    pub fn new(v: usize) -> LocalId {
+    pub fn new(v: u64) -> LocalId {
         LocalId(v)
     }
 }
@@ -58,21 +58,21 @@ impl fmt::Display for LocalId {
 #[derive(Clone, Copy, Debug, Eq, PartialEq, Ord, PartialOrd, Hash, Serialize, Deserialize)]
 pub enum GlobalId {
     /// System namespace.
-    System(usize),
+    System(u64),
     /// User namespace.
-    User(usize),
+    User(u64),
 }
 
 impl GlobalId {
     /// Constructs a new global identifier in the system namespace. It is the
     /// caller's responsibility to provide a unique `v`.
-    pub fn system(v: usize) -> GlobalId {
+    pub fn system(v: u64) -> GlobalId {
         GlobalId::System(v)
     }
 
     /// Constructs a new global identifier in the user namespace. It is the
     /// caller's responsiblity to provide a unique `v`.
-    pub fn user(v: usize) -> GlobalId {
+    pub fn user(v: u64) -> GlobalId {
         GlobalId::User(v)
     }
